HW_1 edited
